Fixes some minor bugs and improves commenting
This update adds a check for if the source or destination directory is
changed. If changed, the all data rest, except for changed value. This
means that if you change the source directory, the destination directory and
all other information is reset and vise versa. However, if you then set the
directory that was reset, only the entry forms are reset.

Commenting on many of the methods has been greatly improved and redundant/
commented out code has been cleaned up.

Also add opening of an internet browser window using the 'Check Website' button
that links to the page where data is to be uploaded.

Waiting on the FTP backend for uploading to the FTP server.

v0.1.8 20181101 2147 CDT