Adds untested FTP upload support
This version has NOT been tested for FTP upload support, but
implementation has started. This is the last piece of the
puzzle.

The ftpUpload class in initialized each time the ftp_upload method
of Meso1819Gui is called. An attempt to upload the files is made,
and if anything goes wrong a Critical dialog window appears.

HAS NOT BEEN TESTED!!!

v0.1.12 20181102 1445 CDT