Updates timing of UTC date check dialog
In previous version, when clicked 'Copy Files', it would ask
if date in correct units, then would check if the date info
was set. This update moves the check for correct units to after
the check of date set check. The dialog for units checking is
alos moved from the Meso1819Gui class to the dateFrame class
in widgets.py.

FTP uploading is fully implemented and works on local test server.

v0.1.14 20181103 1722 CDT