Fixes destination directory when select IOP dir.
When selecting the destination directory, if a user picked
an existing IOP directory, the program appends a second IOP
directory to the file path. This ends up creating
/dst_dir/IOP#/IOP#/date directories. The program now checks
to see if 'IOP' is in the basename of the selected output
directory. If IOP is in the path, then the destination dir
is set to the dirname of the path.

v0.1.15 20181106 1304 CDT