Adds FTP uploading to UCAR and NOAA sites
This update enables uploading of data files to the UCAR
and NOAA data locations for the Meso 18-19 project.

Some other updates include moving warming and confirmation
dialogs to their own file/classes to clean up code in the
Meso1819Gui class.

Also added option to enable debugging logging in the logging
widget using a QCheckBox.

This update also fixes an issue in the iMet2SHARPpy function
that caused the first point in the sounding to be removed.

The loading of sounding files was improved a bit also. Now only
the SPCDecoder is loaded from SHARPpy, which is the only decoder
needed to decode the sounding data.

More comments have been added and code has been cleaned up.

v0.1.17 20181116 1220 CDT