Adds time check before sounding generate
This version adds a time check to see if the current UTC
date/time is before the requested sounding time; i.e., the
sounding is for the future. If the requested date/time IS in
the future, a message is displayed and the 'Generate Sounding'
button is left disabled until the current UTC time is after
the requested time.

This check is implimented because of a 'bug' in the SPC sounding
data file format. The SPC file format uses only a 2-digit year,
and so in the SHARPpy decoder function, when the date/time for
the sounding is in the future; i.e., greater than the current UTC
time, it is assumed the year is 1900. This is the year that then
appears on the sounding image. Thus, the user is forced to wait
so that the correct year is displayed on the sounding image.

v0.1.18 20190204 1605 UTC