Adds indictor lights next to buttons
This update adds indicator lights next to the 'processing' buttons
after they have finished. These were added because depending on
the time requested for the sounding, the generation of the
sounding image may be delayed. The lights give the user some
more feed back on whether a given step has finished. For example,
if the generation of the sounding image is delayed, then a green
light will appear next to the 'Process Files' button to
indicate that the step has finished successfully, even though the
'Generate Sounding' button has not become active yet.

v0.1.19 20190204 1706 UTC