Better checking for file rename/convert
This version adds better checking for file renaming and
conversion in the processing stage of the GUI. Before, if
a file was NOT found for renaming or processing, the user was
not informed; i.e., source file for conversion to SHARPpy format.
It would then be impossible to create the Skew-T without such
files. Now a warning dialog is created and errors logged when
files are not found/processed