Handle missing skew-t file on ftp upload
This version accounts for failure to generate Skew-T when
trying to upload to FTP server. A 'sharppy_bug' flag is added
to the generate sounding method that is set to true on SHARPpy
plotting failure.

When the method then tries to save the skew-t, the sharppy_bug
flag is check. if it is True, then no attempt is made to save the
skew-t and the .png file path is never added to the list of files
to upload